+++ demo/src/bootstrap/index.js
@@ -4,5 +4,5 @@ import App from './app';
 
 ReactDOM.render(
   <App />,
-  document.getElementById('reactGridLayoutWithBuilder')
+  document.getElementById('bootstrapReactGridLayoutBuilder')
 );